Rename openPointInTime and closePointInTime to createPointInTime and deletePointInTime to better match OpenSearch API.
