+++ src/OpenSearch/Client.php
@@ -1246,17 +1246,17 @@ class Client
         return $this->performRequest($endpoint);
     }
     /**
-     * $params['body'] = (array) a point-in-time id to close
+     * $params['body'] = (array) a point-in-time id to delete
      *
      * @param array $params Associative array of parameters
      * @return array
      */
-    public function closePointInTime(array $params = [])
+    public function deletePointInTime(array $params = [])
     {
         $body = $this->extractArgument($params, 'body');
 
         $endpointBuilder = $this->endpoints;
-        $endpoint = $endpointBuilder('ClosePointInTime');
+        $endpoint = $endpointBuilder('DeletePointInTime');
         $endpoint->setParams($params);
         $endpoint->setBody($body);
 
@@ -1273,12 +1273,12 @@ class Client
      * @param array $params Associative array of parameters
      * @return array
      */
-    public function openPointInTime(array $params = [])
+    public function createPointInTime(array $params = [])
     {
         $index = $this->extractArgument($params, 'index');
 
         $endpointBuilder = $this->endpoints;
-        $endpoint = $endpointBuilder('OpenPointInTime');
+        $endpoint = $endpointBuilder('CreatePointInTime');
         $endpoint->setParams($params);
         $endpoint->setIndex($index);
