+++ tests/Endpoints/CreatePointInTimeIntegrationTest.php
@@ -25,12 +25,12 @@ use OpenSearch\Client;
 use OpenSearch\Tests\Utility;
 
 /**
- * Class ClosePointInTimeIntegrationTest
+ * Class OpenPointInTimeIntegrationTest
  *
  * @subpackage Tests\Endpoints
  * @group Integration
  */
-class ClosePointInTimeIntegrationTest extends \PHPUnit\Framework\TestCase
+class CreatePointInTimeIntegrationTest extends \PHPUnit\Framework\TestCase
 {
     private const INDEX = 'movies';
 
@@ -63,55 +63,17 @@ class ClosePointInTimeIntegrationTest extends \PHPUnit\Framework\TestCase
         ]);
     }
 
-    public function testClosePointInTimeSingle()
+    public function testOpenPointInTime()
     {
-        // Arrange
-        $result = $this->client->openPointInTime([
+        // Act
+        $result = $this->client->createPointInTime([
             'index' => self::INDEX,
             'keep_alive' => '10m',
         ]);
-        $pitId = $result['pit_id'];
-
-        // Act
-        $result = $this->client->closePointInTime([
-            'body' => [
-                'pit_id' => $pitId,
-            ],
-        ]);
 
         // Assert
-        $this->assertCount(1, $result['pits']);
-        $pit = $result['pits'][0];
-        $this->assertTrue($pit['successful']);
-        $this->assertSame($pitId, $pit['pit_id']);
-    }
-
-    public function testClosePointInTimeMultiple()
-    {
-        // Arrange
-        $result = $this->client->openPointInTime([
-            'index' => self::INDEX,
-            'keep_alive' => '10m',
-        ]);
-        $pitId1 = $result['pit_id'];
-        $result = $this->client->openPointInTime([
-            'index' => self::INDEX,
-            'keep_alive' => '1h',
-        ]);
-        $pitId2 = $result['pit_id'];
-
-        // Act
-        $result = $this->client->closePointInTime([
-            'body' => [
-                'pit_id' => [$pitId1, $pitId2],
-            ],
-        ]);
-
-        // Assert
-        $this->assertCount(2, $result['pits']);
-        foreach ($result['pits'] as $pit) {
-            $this->assertTrue($pit['successful']);
-            $this->assertTrue(in_array($pit['pit_id'], [$pitId1, $pitId2]));
-        }
+        $this->assertNotEmpty($result['pit_id']);
+        $this->assertGreaterThanOrEqual(time() * 1000, $result['creation_time']);
+        $this->assertLessThan((time() + 10) * 1000, $result['creation_time']);
     }
 }
